serveur (production in progress)

+++ convert_to_nwb_for_AO.py
@@ -1,10 +1,15 @@
 """_summary_
 """
-import importlib
 import os
 import h5py
-from pynwb import NWBHDF5IO, validate
+import gc
+import sys
+import importlib
 import argparse
+import platform
+from tqdm import tqdm
+from pynwb import NWBHDF5IO, validate
+from contextlib import redirect_stdout
 import converters.behavior_to_nwb
 import converters.nwb_saving
 import converters.general_to_nwb
@@ -20,7 +25,7 @@ import converters.intervals_to_nwb
 #############################################################
 
 
-def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth_bin=0.010 ):
+def convert_data_to_nwb_an_mat(mat_file, output_folder):
     """
     Converts data from a config file to an NWB file.
     :param config_file: Path to the yaml config file containing mouse ID and metadata for the session to convert
@@ -36,7 +41,7 @@ def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth
         data = {key: data_group[key][()] for key in data_group.keys()}
 
     print("**************************************************************************")
-    print("-_-_-_-_-_-_-_-_-_-_-_-_-_-_- NWB conversion _-_-_-_-_-_-_-_-_-_-_-_-_-_-_")
+    print(f" NWB conversion {mat_file}")
     print(" ")
     print(f"📃 Creating config file for NWB conversion :")
     importlib.reload(converters.Initiation_nwb)
@@ -60,15 +65,16 @@ def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth
     print("         - Device metadata")
     print("         - Extracellular electrophysiology metadata")
     
-    print("     o 📶 Add acquisition container")
-    converters.acquisition_to_nwb.add_lfp_acquisition(nwb_file=nwb_file, signal_array=signal, electrode_region=electrode_table_region) # same for rewarded and non-rewarded sessions  
+    #print("     o 📶 Add acquisition container")
+    #converters.acquisition_to_nwb.add_lfp_acquisition(nwb_file=nwb_file, signal_array=signal, electrode_region=electrode_table_region) # same for rewarded and non-rewarded sessions  
 
     print("     o ⏸️ Add intervall container")
     importlib.reload(converters.intervals_to_nwb)
     if Rewarded:
+        u=0
         converters.intervals_to_nwb.add_intervals_container_Rewarded(nwb_file=nwb_file, data=data, mat_file=mat_file)
-    #else:
-        #converters.intervals_to_nwb.add_intervals_container_NonRewarded(nwb_file=nwb_file, data=data, mat_file=mat_file)
+    else:
+        converters.intervals_to_nwb.add_intervals_container_NonRewarded(nwb_file=nwb_file, data=data, mat_file=mat_file)
 
     print("     o 🧠 Add units container")
     importlib.reload(converters.units_to_nwb)
@@ -86,8 +92,8 @@ def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth
         converters.behavior_to_nwb.add_behavior_container_NonRewarded(nwb_file=nwb_file, data=data, config_file=config_file)
 
     print("         - No ephys data for AN sessions")
-    print("         - Analysis complementary information")
-    converters.analysis_to_nwb.add_analysis_container(nwb_file=nwb_file, Rewarded=Rewarded, psth_window=psth_window, psth_bin=psth_bin) # almost same for rewarded and non-rewarded sessions
+    #print("         - Analysis complementary information")
+    #converters.analysis_to_nwb.add_analysis_container(nwb_file=nwb_file, Rewarded=Rewarded, psth_window=psth_window, psth_bin=psth_bin) # almost same for rewarded and non-rewarded sessions
 
     importlib.reload(converters.nwb_saving)
     nwb_path = converters.nwb_saving.save_nwb_file(nwb_file=nwb_file, output_folder=output_folder) # same for rewarded and non-rewarded sessions
@@ -114,10 +120,51 @@ def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth
     print("**************************************************************************")
 
     # Delete .yaml config file 
-    if os.path.exists(output_path):
+    if os.path.exists(output_path) and output_path.endswith('.yaml'):
         os.remove(output_path)
 
 
+def convert_data_to_nwb_an(input_folder, output_folder,print_progress=False):
+    """
+    Converts all .mat files in a folder to NWB format for AN sessions.
+    :param input_folder: Path to the folder containing .mat files
+    :param output_folder: Path to the folder where NWB files will be saved
+    """
+    assert os.path.exists(input_folder), "file input_folder does not exist"
+    files = [f for f in os.listdir(input_folder) if f.endswith(".mat")]
+    # Iterate over each .mat file in the folder
+    list_errors_files = []
+    list_errors = []
+    i = 0
+    with tqdm(files, desc="Conversion .mat files") as pbar:
+        for file in pbar:
+            i += 1
+            full_path = os.path.join(input_folder, file)
+            pbar.set_description(f"Conversion to NWB: 🔁 {file}")
+            try:
+                if print_progress:
+                    convert_data_to_nwb_an_mat(mat_file=full_path,output_folder=output_folder)
+                    def clear_console():
+                        if platform.system() == "Windows":
+                            os.system("cls")
+                        else:
+                            os.system("clear")
+                    clear_console()
+                else:
+                    with open(os.devnull, "w") as fnull, redirect_stdout(fnull):
+                        convert_data_to_nwb_an_mat(mat_file=full_path,output_folder=output_folder)
+            except Exception as e:
+                #print(f"⚠️ Error in {file} : {e}")
+                list_errors_files.append(file)
+                list_errors.append(str(e))
+            if i == len(files):
+                pbar.set_description(f"Conversion to NWB is finished")
+        if len(list_errors_files) > 0:
+            print(f"⚠️ Conversion completed with errors for {len(list_errors_files)} files")
+            for i, file in enumerate(list_errors_files):
+                print(f"    - {file}: {list_errors[i]}")
+        gc.collect() 
+    return None
 
 #_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_
 #_-_-_-_-_-_-_-_-_-_-_-_-_-_-_ MAIN _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
@@ -125,16 +172,10 @@ def convert_data_to_nwb_an(mat_file, output_folder,psth_window=(-0.2, 0.5), psth
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Convert data to NWB format for AN sessions")
-    parser.add_argument("mat_file", type=str, help="Path to the .mat file containing the data")
-    parser.add_argument("output_folder", type=str, help="Path to the folder where the NWB file will be saved")
-    parser.add_argument("--psth_window", type=float, nargs=2, default=(-0.2, 0.5), help="PSTH and LFP window in seconds for analysis")
-    parser.add_argument("--psth_bin", type=float, default=0.010, help="PSTH and LFP bin size in seconds for analysis")
+    parser.add_argument("input_folder", type=str, help="Path to the folder containing .mat files")
+    parser.add_argument("output_folder", type=str, help="Path to the folder where the NWB files will be saved")
+    parser.add_argument("--print_progress", action="store_true", help="Print progress of conversion")
+
 
     args = parser.parse_args()
-
-    convert_data_to_nwb_an(
-        mat_file=args.mat_file,
-        output_folder=args.output_folder,
-        psth_window=tuple(args.psth_window),
-        psth_bin=args.psth_bin
-    )
+    convert_data_to_nwb_an(input_folder=args.input_folder, output_folder=args.output_folder , print_progress=args.print_progress)
